MueLu: fix compile warnings in tests
Build/Test Cases Summary
Enabled Packages: MueLu
Disabled Packages: PyTrilinos,Claps,TriKota
Enabled all Forward Packages
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=121,notpassed=0 (69.36 min)
Other local commits for this build/test group: 0574ea1

+++ packages/muelu/test/unit_tests/Smoothers/BlockedDirectSolver.cpp
@@ -436,7 +436,7 @@ namespace MueLuTests {
       solver->Apply(*X, *RHS, true);  //zero initial guess
       Teuchos::ArrayRCP<const Scalar> xdata = X->getData(0);
       bool bCheck = true;
-      for(int i=0; i<X->getLocalLength(); i++) {
+      for(size_t i=0; i<X->getLocalLength(); i++) {
         if (i>=0  && i< 5) { if(xdata[i] != (SC) 1.0) bCheck = false; }
         if (i>=5 && i< 10) { if(xdata[i] != (SC) 1.0/2.0) bCheck = false; }
         if (i>=10 && i< 20) { if(xdata[i] != (SC) 1.0/3.0) bCheck = false; }
@@ -546,7 +546,7 @@ namespace MueLuTests {
       solver->Apply(*X, *RHS, true);  //zero initial guess
       Teuchos::ArrayRCP<const Scalar> xdata = X->getData(0);
       bool bCheck = true;
-      for(int i=0; i<X->getLocalLength(); i++) {
+      for(size_t i=0; i<X->getLocalLength(); i++) {
         if (i>=0  && i< 5) { if(xdata[i] != (SC) 1.0) bCheck = false; }
         if (i>=5 && i< 10) { if(xdata[i] != (SC) 1.0/2.0) bCheck = false; }
         if (i>=10 && i< 20) { if(xdata[i] != (SC) 1.0/3.0) bCheck = false; }
@@ -714,7 +714,7 @@ namespace MueLuTests {
       simpleSmoother->Apply(*X, *RHS, true);  //zero initial guess
       Teuchos::ArrayRCP<const Scalar> xdata = X->getData(0);
       bool bCheck = true;
-      for(int i=0; i<X->getLocalLength(); i++) {
+      for(size_t i=0; i<X->getLocalLength(); i++) {
         if (i>=0  && i< 10) { if(xdata[i] != (SC) 1.0/3.0) bCheck = false; }
         if (i>=10 && i< 15) { if(xdata[i] != (SC) 1.0) bCheck = false; }
         if (i>=15 && i< 20) { if(xdata[i] != (SC) 0.5) bCheck = false; }
@@ -881,7 +881,7 @@ namespace MueLuTests {
       simpleSmoother->Apply(*X, *RHS, true);  //zero initial guess
       Teuchos::ArrayRCP<const Scalar> xdata = X->getData(0);
       bool bCheck = true;
-      for(int i=0; i<X->getLocalLength(); i++) {
+      for(size_t i=0; i<X->getLocalLength(); i++) {
         if (i>=0  && i< 10) { if(xdata[i] != (SC) 1.0/3.0) bCheck = false; }
         if (i>=10 && i< 15) { if(xdata[i] != (SC) 1.0) bCheck = false; }
         if (i>=15 && i< 20) { if(xdata[i] != (SC) 0.5) bCheck = false; }
